I am trying to do some more on the Board class, and have run into the problem of checking if a piece is playable or not.
My idea is to create a board, then have a method that check if the adjacent cases of a given case is playable or not. It should check if the adjacent cases is occupied or not, and if not, compare the value of the current piece with the pieces on the board. We can discuss how to deal with this when we have time this week.